Better organized CSV fields and objects to reflect that.
