+++ src/puzzles.csv
@@ -1,3 +1,3 @@
-DIFFICULTY,PUZZLE,SOLUTION
-Undefined,003000000600403000008200316070180000064302190000056030916005800000601005000000400,193568247627413589458297316379184652564372198281956734916745823842631975735829461
-Undefined,860050300250800640000000000046570800009010400003064250000000000038007024005080076,864759312257831649391426785146572893529318467783964251972645138638197524415283976
+IDX,DIFFICULTY,PUZZLE,IS_STARTED,CURRENT,SOLUTION
+0,UNDEFINED,003000000600403000008200316070180000064302190000056030916005800000601005000000400,FALSE,003000000600403000008200316070180000064302190000056030916005800000601005000000400,193568247627413589458297316379184652564372198281956734916745823842631975735829461
+1,UNDEFINED,860050300250800640000000000046570800009010400003064250000000000038007024005080076,FALSE,860050300250800640000000000046570800009010400003064250000000000038007024005080076,864759312257831649391426785146572893529318467783964251972645138638197524415283976
